requirements_pc.txt
pygame
numpy
matplotlib
pandas
pillow
scikit-learn
pyzmq